+++ src/main/java/io/leangen/graphql/metadata/execution/Executable.java
@@ -1,21 +1,16 @@
 package io.leangen.graphql.metadata.execution;
 
-import io.leangen.graphql.metadata.exceptions.MappingException;
-
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
 import java.lang.reflect.AnnotatedElement;
 import java.lang.reflect.AnnotatedType;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Member;
-import java.lang.reflect.Modifier;
 import java.lang.reflect.Parameter;
 
 public abstract class Executable<T extends AnnotatedElement & Member> {
 
     T delegate;
-    MethodHandle handle;
 
-    abstract public Object execute(Object target, Object[] args) throws Throwable;
+    abstract public Object execute(Object target, Object[] args) throws InvocationTargetException, IllegalAccessException;
 
     abstract public AnnotatedType getReturnType();
 
@@ -35,22 +30,6 @@ public abstract class Executable<T extends AnnotatedElement & Member> {
         return delegate;
     }
 
-    MethodHandle prepareHandle(MethodHandleProvider provider) {
-        try {
-            MethodHandle handle = provider.get();
-            if (Modifier.isStatic(delegate.getModifiers())) {
-                handle = MethodHandles.dropArguments(handle, 0, Object.class);
-            }
-            return handle.asFixedArity();
-        } catch (IllegalAccessException e) {
-            throw new MappingException("Method '" + delegate.getName() + "' can not be mapped", e);
-        }
-    }
-
-    interface MethodHandleProvider {
-        MethodHandle get() throws IllegalAccessException;
-    }
-
     @Override
     public int hashCode() {
         return delegate.hashCode();
@@ -63,8 +42,8 @@ public abstract class Executable<T extends AnnotatedElement & Member> {
      * {@inheritDoc}
      */
     @Override
-    public boolean equals(Object other) {
-        return other instanceof Executable && ((Executable) other).delegate.equals(this.delegate);
+    public boolean equals(Object that) {
+        return this == that || (that instanceof Executable && ((Executable) that).delegate.equals(this.delegate));
     }
 
     @Override